SEO Tools: Fix upgrade nudge for Jetpack sites
The new Jetpack plans didn't include the SEO preview tools feature, and
so we would display a nudge to upgrade to a business plan. Looking at
the legacy plans, it seems as though this feature was available on all
paid Jetpack plans, not just business.

Testing with a premium WPCOM plan, the meta section isn't shown, which
is the same for free Jetpack sites. Therefore it would appear that we
don't display this nudge in this context.

This change adds the feature to the new Jetpack plans and updates the
nudge to prompt people with Jetpack sites to upgrade to Security Daily.